Add item requests

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -2,6 +2,8 @@ package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.repository.BookingRepository;
@@ -33,9 +35,12 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<Item> getAllItemsByOwnerId(Long ownerId) {
-        List<Item> itemList = itemRepository.findByOwnerId(ownerId);
-        itemList.sort(Comparator.comparing(Item::getId));
+    public List<Item> getAllItemsByOwnerId(Long ownerId, Integer from, Integer size) {
+        Pageable pageable = PageRequest.of(from / size, size);
+        List<Item> itemList = itemRepository.findByOwnerId(ownerId, pageable);
+        if (itemList.size() > 1) {
+            itemList.sort(Comparator.comparing(Item::getId));
+        }
         return itemList;
     }
 
@@ -65,11 +70,12 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<Item> searchItemsByTextInNameAndDescription(String text) {
+    public List<Item> searchItemsByTextInNameAndDescription(String text, Integer from, Integer size) {
         if (text.isBlank()) {
             return Collections.emptyList();
         }
-        return itemRepository.searchItemsByTextInNameAndDescription(text);
+        Pageable pageable = PageRequest.of(from / size, size);
+        return itemRepository.searchItemsByTextInNameAndDescription(text, pageable);
     }
 
     @Override
@@ -89,7 +95,16 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public void deleteItem(Long itemId) {
+    public List<Item> getItemsByRequestId(Long requestId) {
+        return itemRepository.findItemsByRequestId(requestId);
+    }
+
+    @Override
+    public void deleteItemById(Long ownerId, Long itemId) {
+        Item item = getItemById(itemId);
+        if (!item.getOwner().getId().equals(ownerId)) {
+            throw new ValidationException("Удалить вещь может только её владелец");
+        }
         itemRepository.deleteById(itemId);
     }
 }
